[core/tx-runner] split across multiple files
