+++ src/renderer/coreapp/app.tsx
@@ -13,7 +13,8 @@ import AccountsBar from './accounts-bar'
 import AppMenu from './app-menu'
 import CheckUpdate from './check-update'
 import AccountsApp from './accounts-app/accounts-app'
-import TXRunner, { TXCancelled } from './tx-runner/tx-runner'
+import TXRunner from './tx-runner/tx-runner'
+import { TXCancelled } from './tx-runner/run-txs'
 
 import { AppList } from '../apps/apps'
 import Accounts from './accounts-app/def'
